+++ test/runtests.jl
@@ -13,10 +13,11 @@ using ExplicitImports
 
 (isinteractive() ? includet : include)(joinpath(pkgdir(NetworkDynamics, "test", "testutils.jl")))
 
-haskey(ENV, "BUILDKITE") && @test CUDA.functional() # fail early in buildkite if cuda is not available
+BUILDKITE = haskey(ENV, "BUILDKITE")
+BUILDKITE && @test CUDA.functional() # fail early in buildkite if cuda is not available
 
 @testset "NetworkDynamics Tests" begin
-    @testset "Package Quality Tests" begin
+    BUILDKITE || @testset "Package Quality Tests" begin
         # print_explicit_imports(NetworkDynamics)
         @test check_no_implicit_imports(NetworkDynamics) === nothing
         @test check_no_stale_explicit_imports(NetworkDynamics, ignore=(:Symbolics,)) === nothing
@@ -27,41 +28,44 @@ haskey(ENV, "BUILDKITE") && @test CUDA.functional() # fail early in buildkite if
         @test_broken isempty(Docs.undocumented_names(NetworkDynamics))
     end
 
-    @safetestset "utils test" begin include("utils_test.jl") end
+    # skip non-GPU tests on buildkite
+    if !BUILDKITE
+        @safetestset "utils test" begin include("utils_test.jl") end
 
-    NetworkDynamics.CHECK_COMPONENT[] = false
-    @safetestset "construction test" begin include("construction_test.jl") end
-    @safetestset "Aggregation Tests" begin include("aggregators_test.jl") end
-    @safetestset "massmatrix test" begin include("massmatrix_test.jl") end
-    NetworkDynamics.CHECK_COMPONENT[] = true
+        NetworkDynamics.CHECK_COMPONENT[] = false
+        @safetestset "construction test" begin include("construction_test.jl") end
+        @safetestset "Aggregation Tests" begin include("aggregators_test.jl") end
+        @safetestset "massmatrix test" begin include("massmatrix_test.jl") end
+        NetworkDynamics.CHECK_COMPONENT[] = true
+
+        @safetestset "initialization test" begin include("initialization_test.jl") end
+        @safetestset "Callbacks test" begin include("callbacks_test.jl") end
+        @safetestset "Metadata test" begin include("metadata_test.jl") end
+        @safetestset "Linear Stability test" begin include("linear_stability_test.jl") end
+        @safetestset "Show-methods test" begin include("show_test.jl") end
+        @safetestset "Spinners test" begin include("spinners_test.jl") end
+        @safetestset "sparsity test" begin include("sparsity_test.jl") end
+
+        @safetestset "MTK extension test" begin include("MTK_test.jl") end
+
+        # check on the precompile files
+        @safetestset "Precompile workload" begin include("../src/precompile_workload.jl") end
+        @safetestset "MTK precompile workload" begin include("../ext/MTKExt_precomp_workload.jl") end
+
+        @safetestset "AD test" begin include("AD_test.jl") end
+        @safetestset "doctor test" begin include("doctor_test.jl") end
+    end
 
     @safetestset "Symbolic Indexing Tests" begin include("symbolicindexing_test.jl") end
     @safetestset "external input test" begin include("external_inputs_test.jl") end
 
-    @safetestset "doctor test" begin include("doctor_test.jl") end
 
     @safetestset "Diffusion test" begin include("diffusion_test.jl") end
     @safetestset "inhomogeneous test" begin include("inhomogeneous_test.jl") end
-    @safetestset "initialization test" begin include("initialization_test.jl") end
-    @safetestset "Callbacks test" begin include("callbacks_test.jl") end
-    @safetestset "Metadata test" begin include("metadata_test.jl") end
-    @safetestset "Linear Stability test" begin include("linear_stability_test.jl") end
-    @safetestset "Show-methods test" begin include("show_test.jl") end
-    @safetestset "Spinners test" begin include("spinners_test.jl") end
-
-    @safetestset "sparsity test" begin include("sparsity_test.jl") end
 
     if CUDA.functional()
         @safetestset "GPU test" begin include("GPU_test.jl") end
     end
-
-    @safetestset "MTK extension test" begin include("MTK_test.jl") end
-
-    # check on the precompile files
-    @safetestset "Precompile workload" begin include("../src/precompile_workload.jl") end
-    @safetestset "MTK precompile workload" begin include("../ext/MTKExt_precomp_workload.jl") end
-
-    @safetestset "AD test" begin include("AD_test.jl") end
 end
 
 @testset "Test Doc Examples" begin